--- src/main/java/com/sejong/userservice/api/controller/dto/JoinRequest.java
@@ -1,38 +1,0 @@
-package com.sejong.userservice.api.controller.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class JoinRequest {
-    @NotBlank(message = "Username cannot be empty")
-    private String username;
-
-    @NotBlank(message = "Password cannot be empty")
-    @Size(min = 8, message = "Password must be at least 8 characters long")
-    private String password;
-
-    @NotBlank(message = "Real name cannot be empty")
-    private String realName;
-
-    @NotBlank(message = "Email cannot be empty")
-    @Email(message = "Invalid email format")
-    private String email;
-
-    @Min(value = 1, message = "Grade must be at least 1")
-    @Max(value = 5, message = "Grade cannot be more than 4")
-    private Integer grade;
-
-    @NotBlank(message = "Major cannot be empty")
-    private String major;
-}
