--- src/main/java/com/sejong/userservice/api/controller/dto/UserResponse.java
@@ -1,45 +1,0 @@
-package com.sejong.userservice.api.controller.dto;
-
-import com.sejong.userservice.domain.model.User;
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UserResponse {
-    private Long id;
-    private String username;
-    private String role;
-    private String realName;
-    private String email;
-    private Integer grade;
-    private String major;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-
-    public static UserResponse from(User user) {
-        if (user == null) {
-            return null;
-        }
-        return UserResponse.builder()
-                .id(user.getId())
-                .username(user.getUsername())
-                .role(user.getRole())
-                .realName(user.getRealName())
-                .email(user.getEmail())
-                .grade(user.getGrade())
-                .major(user.getMajor())
-                .createdAt(user.getCreatedAt())
-                .updatedAt(user.getUpdatedAt())
-                .build();
-    }
-}
-
-
