+++ src/main/java/com/sejong/userservice/application/exception/TokenReissueException.java
@@ -1,0 +1,14 @@
+package com.sejong.userservice.application.exception;
+
+import com.sejong.userservice.application.token.TokenReissueStatus;
+import lombok.Getter;
+
+@Getter
+public class TokenReissueException extends RuntimeException {
+    private final TokenReissueStatus status;
+
+    public TokenReissueException(String message, TokenReissueStatus status) {
+        super(message);
+        this.status = status;
+    }
+}
