--- src/main/java/com/sejong/userservice/application/service/token/TokenReissueException.java
@@ -1,13 +1,0 @@
-package com.sejong.userservice.application.service.token;
-
-import lombok.Getter;
-
-@Getter
-public class TokenReissueException extends RuntimeException {
-    private final TokenReissueStatus status;
-
-    public TokenReissueException(String message, TokenReissueStatus status) {
-        super(message);
-        this.status = status;
-    }
-}
