+++ src/main/java/com/sejong/userservice/application/user/CustomUserDetailsService.java
@@ -1,0 +1,31 @@
+package com.sejong.userservice.application.user;
+
+import com.sejong.userservice.application.user.dto.CustomUserDetails;
+import com.sejong.userservice.core.user.User;
+import com.sejong.userservice.core.user.User.UserRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    public CustomUserDetailsService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = userRepository.findByUsername(username);
+
+        if (user != null) {
+            return new CustomUserDetails(user);
+        }
+
+        throw new UsernameNotFoundException("엥 그런 유저 네임은 없어요!!!");
+    }
+
+}
