--- src/main/java/com/sejong/userservice/domain/repository/RefreshTokenRepository.java
@@ -1,19 +1,0 @@
-package com.sejong.userservice.domain.repository;
-
-import java.time.LocalDateTime;
-
-public interface RefreshTokenRepository {
-    String saveRefreshToken(String token, String username, LocalDateTime expiryDate, String jti);
-
-    String findUsernameByToken(String token);
-
-    String findUsernameByJti(String jti);
-
-    boolean revokeToken(String token);
-
-    boolean revokeTokenByJti(String jti);
-
-    boolean isTokenValidOnServer(String token);
-
-    void revokeAllTokensForUser(String username);
-}
