--- src/main/java/com/sejong/userservice/exception/DuplicateUsernameException.java
@@ -1,16 +1,0 @@
-package com.sejong.userservice.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.CONFLICT)
-public class DuplicateUsernameException extends RuntimeException {
-
-    public DuplicateUsernameException(String message) {
-        super(message);
-    }
-
-    public DuplicateUsernameException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
